pyproject.toml
[project]
name = "machineminds-trailmate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "googlemaps>=4.10.0",
    "groq>=0.30.0",
    "openai>=1.95.1",
    "openai-agents[litellm]>=0.1.0",
    "playwright>=1.53.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
